Encrypt data access password

+++ rentmanager/companyfilepassword.cpp
@@ -2,6 +2,7 @@
 #include "ui_companyfilepassword.h"
 #include <QMessageBox>
 #include <QtSql>
+#include "simplecrypt.h"
 CompanyFilePassword::CompanyFilePassword(QWidget *parent, bool newP) :
 	QDialog(parent),
 	ui(new Ui::CompanyFilePassword)
@@ -20,28 +21,31 @@ CompanyFilePassword::~CompanyFilePassword()
 
 void CompanyFilePassword::on_cmdOK_clicked()
 {
-    if (ui->txtConfirmPassword->text() != ui->txtNewPassword->text()) {
-	    QMessageBox::critical(this, "Check password", "Your passwords do not match. Please re-enter");
-	    return;
-    }
-
-    if (!isNewP) {
-	QSqlQuery qu = QSqlDatabase::database().exec("SELECT * FROM password");
-	if (!qu.lastError().isValid()) {
-		qu.first();
-		QString oldP = qu.record().value("Password").toString();
-		if (oldP != ui->txtCurrentPassword->text()) {
-			QMessageBox::critical(this, "Check password", "Your current password is incorrect. Please re-enter");
-			return;
-		}
+	if (ui->txtConfirmPassword->text() != ui->txtNewPassword->text()) {
+		QMessageBox::critical(this, "Check password", "Your passwords do not match. Please re-enter");
+		return;
 	}
-	QSqlDatabase::database().exec("CREATE TABLE IF NOT EXISTS 'password' ('Password' TEXT)");
-	QSqlDatabase::database().exec("DELETE FROM password");
-	QSqlDatabase::database().exec("INSERT INTO password ('Password') VALUES ('" + ui->txtNewPassword->text() + "')");
-	this->accept();
-    }
+	SimpleCrypt crypt(Q_UINT64_C(0x0c2ad4a4acb9f023));
+	QString encryptedPassword = crypt.encryptToString(ui->txtNewPassword->text());
+	QString oldEncryptedPassword = crypt.encryptToString(ui->txtCurrentPassword->text());
+	if (!isNewP) {
+		QSqlQuery qu = QSqlDatabase::database().exec("SELECT * FROM password");
+		if (!qu.lastError().isValid()) {
+			qu.first();
+			QString oldP = qu.record().value("Password").toString();
+			oldP = crypt.decryptToString(oldP);
+			if (oldP != ui->txtCurrentPassword->text()) {
+				QMessageBox::critical(this, "Check password", "Your current password is incorrect. Please re-enter");
+				return;
+			}
+		}
+		QSqlDatabase::database().exec("CREATE TABLE IF NOT EXISTS 'password' ('Password' TEXT)");
+		QSqlDatabase::database().exec("DELETE FROM password");
+		QSqlDatabase::database().exec("INSERT INTO password ('Password') VALUES ('" + encryptedPassword + "')");
+		this->accept();
+	}
 
-    newPassword = ui->txtNewPassword->text();
-    currentPassword = ui->txtCurrentPassword->text();
-    this->accept();
+	newPassword = encryptedPassword;
+	currentPassword = oldEncryptedPassword;
+	this->accept();
 }
